PHP8 support: Starting point for getting it to compile
PHP8 requires at least SWIG 4.1.0 - we use SWIG 4.1.1 in this commit.
Still we get a number of errors from both PHP AIP, which has been attempted to
be corrected in this branch. But the SWIG generated wrapper still has compiler and
linker errors.
So any help is really apreciated a lot :)
Feel free to contact me.